testing script and QC script put into functions
